dockerHub service connection and docker-compose run integration test services
Added two tasks: 1. A task to log into dockerHub so that custom images for the integration test can be retrieved. 2. A task to start services from the docker-compose file which uses those images. The idea is that those services can then be reached from the automated tests run by Maven in the next step.